Allow enforcing of whole /etc/hosts
Parameter purge_hosts will enforce whole /etc/hosts file, removing entries
that are not defined in model except defaults for both IPv4 and IPv6 localhost
and hostname + fqdn.

It's good to use this option if you want to ensure /etc/hosts is always in a
clean state however it's not enabled by default for safety.

Change-Id: Ieb219b2dd494d5a66ecc52a2ec00dd5157ee2b7a