Fix ordering between repo config and package install
In cases of having:
`  linux:
     system:
       purge_repos: true`
in reclass, and using local repos, we'll get the following issue:
` - E: Unable to locate package libmnl0`
because we do not have repositories configured yet and do not have
direct access to the internet.
In sush case we need to revert ordering, configure repo -> install
package.