Use Pillar to chose which interfaces to monitor.
The `linux_netlink.ls` function used a regex to choose which interfaces
to collect metric for.

`_alphanum_re = re.compile(r'^[a-z0-9]+$')`

Unfortunately, by default this excludes vlan and tap interfaces, which
are kind of important.  ie `bond0.120` or `tap2a3dab86-fb`.

We also have a problem where even if we update the regex to include
these interfaces... if someone deletes and spawns a new instance then
the tap device name changes on the compute host, which will not be
monitored unless someone re-runs the `collectd` on the compute again.
Less than ideal.

This commit lets us choose `VerboseInterface "all"` using Pillar data
to avoid this problem.