Invert skip_verify to be opposite of "verfy" flag in pillar
- skip_verify: {{ true if package.verify else false }}
will be true when package:verify is set to true
And that confuses.

"skip verify" means "skip any verify process" and package should
not be verified during installation

Currently, for disabling verification we use:
linux:
  system:
    package:
      netapp-offload-tool:
        verify: true

And anyone who is reading reclass model can be easily perplexed.
Patchset fixes that issue.

Change-Id: I38d60ff3dbe651e14898c4937fc7dddad36204be