system.repo: Debian: Use proxy for keyservers
Previously, when fetching GPG keys for APT keyring, either using
public key download & import (as for default repos) or via keyserver,
we relied on simple `curl` calls or passed it down to Salt aptpkg
module.
To be able to retrieve APT keys behind a proxy, one used to have to
configure the proxy for the Salt minion, which does not yet have
`no_proxy` support (either *all* or *no* traffic hits the proxy).

When `linux:system:proxy` http(s) proxies are set:
- no longer pass key configuration to Salt aptpkg (until it properly
  supports `no_proxy`);
- handle all keys explicitly with `curl` and `apt-key`;
- set 'http(s)_proxy' env vars for `cmd.wait` calls;

If `linux:system:proxy` is not defined, the behavior is
unchanged for backwards compatibility.

NOTE: If present, per-repo proxies are also used for keyserver access.

system.repo: Fix conditions order for Debian proxy

Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>