Network resolf.conf handling
the handling as the Resolv.conf is generated and adapted, adapted.
previously the Resolv.conf was created and then through
Overwrite "network.system" in the interface.sls again.
With two search servers that should actually be included.
"search example.com. sudomain.example.com"
but it always became that
search ['example.com.', 'sudomain.example.com']

The resolv.conf was first created correctly but then overwritten again in the interface.sls.

The problem only arises if you don't want to have a "Domain:" in resov.conf